--- noriter/src/test/java/com/codewarts/noriter/article/unit/service/StudyServiceTest.java
@@ -1,85 +1,0 @@
-package com.codewarts.noriter.article.unit.service;
-
-import static io.restassured.RestAssured.given;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.HttpHeaders.CONNECTION;
-import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
-import static org.springframework.http.HttpHeaders.DATE;
-import static org.springframework.http.HttpHeaders.HOST;
-import static org.springframework.http.HttpHeaders.TRANSFER_ENCODING;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
-
-import com.codewarts.noriter.auth.jwt.JwtProvider;
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.specification.RequestSpecification;
-import java.util.List;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-
-@DisplayNameGeneration(ReplaceUnderscores.class)
-@ExtendWith({RestDocumentationExtension.class})
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class StudyServiceTest {
-
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    protected JwtProvider jwtProvider;
-
-    protected RequestSpecification documentationSpec;
-
-    @BeforeEach
-    void setup(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-        documentationSpec = new RequestSpecBuilder()
-            .addFilter(
-                documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(
-                        prettyPrint(),
-                        removeHeaders(HOST, CONTENT_LENGTH))
-                    .withResponseDefaults(
-                        prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING,
-                            "Keep-Alive",
-                            HttpHeaders.VARY))
-            )
-            .build();
-    }
-
-    @Test
-    void 글을_등록한다() {
-        String accessToken = jwtProvider.issueAccessToken(2L);
-
-        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "content", "헬륨가스를모곳지", "hashtags", List.of("SPRING", "JPA", "코린이"));
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .body(requestBody)
-
-            .when()
-            .post("/community/gathering")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-}
